add djwip.ixcat.djgrid: jupyter data entry widget.
main usage is shown in djgrid.ipynb.

To note, this is currently implemented via qgrid,
which in turn is designed to manipulate pandas dataframes -
as such it is a bit cumbersome and non native, requiring converting
dj tables 1st to pandas frames, then editing, then saving.

Additionally:

- The pandas frame currently generates a synthetic 'index' number
- Editing only implemented for 'new' rows

Goal is to:

1) Get round trip fetch/edit/save going in a single user call
2) Fixup the 'update' method to support arbitrary edits
3) Having a stable UI after #1, attempt to bypass the pandas
   dependency and natively operate on datasets.